move component out too its own package
