+++ libs/design/select/examples/ng-package.json
@@ -4,6 +4,6 @@
   "deleteDestPath": false,
   "lib": {
     "entryFile": "src/index.ts",
-    "styleIncludePaths": ["../../scss"]
+    "styleIncludePaths": ["../../src/scss"]
   }
 }
