--- libs/design/select/examples/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "@daffodil/design/select/examples"
-}
