+++ libs/design/select/examples/src/default-select/default-select.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { DaffSelectModule } from '@daffodil/design';
+import { DaffSelectModule } from '@daffodil/design/select';
 
 import { DefaultSelectComponent } from './default-select.component';
