+++ libs/design/select/examples/src/disabled-select/disabled-select.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { DaffSelectModule } from '@daffodil/design';
+import { DaffSelectModule } from '@daffodil/design/select';
 
 import { DisabledSelectComponent } from './disabled-select.component';
