+++ libs/design/select/examples/src/select-with-error/select-with-error.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { DaffSelectModule } from '@daffodil/design';
+import { DaffSelectModule } from '@daffodil/design/select';
 
 import { SelectWithErrorComponent } from './select-with-error.component';
