+++ libs/design/select/src/animation/select-animation-state.spec.ts
