+++ libs/design/select/src/animation/select-animation-state.ts
