+++ libs/design/select/src/animation/select-animation.ts
