+++ libs/design/select/src/index.ts
@@ -1,0 +1,1 @@
+export * from './public_api';
