+++ libs/design/select/src/option/option.directive.ts
