+++ libs/design/select/src/public_api.ts
@@ -1,4 +1,4 @@
 export { DaffSelectModule } from './select.module';
-export * from './select.component';
+export * from './select/select.component';
 export * from './option/option.directive';
 export * from './option/context.type';
