+++ libs/design/select/src/select.module.ts
@@ -5,11 +5,14 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
+import {
+  DaffErrorMessageModule,
+  DaffFormLabelModule,
+  DaffPrefixSuffixModule,
+} from '@daffodil/design';
+
 import { DaffSelectOptionDirective } from './option/option.directive';
-import { DaffSelectComponent } from './select.component';
-import { DaffErrorMessageModule } from '../../atoms/form/error-message/error-message.module';
-import { DaffFormLabelModule } from '../../atoms/form/form-label/form-label.module';
-import { DaffPrefixSuffixModule } from '../../core/prefix-suffix/prefix-suffix.module';
+import { DaffSelectComponent } from './select/select.component';
 
 @NgModule({
   imports: [
@@ -17,10 +20,11 @@ import { DaffPrefixSuffixModule } from '../../core/prefix-suffix/prefix-suffix.m
     FontAwesomeModule,
     OverlayModule,
     PortalModule,
+    ReactiveFormsModule,
+
     DaffPrefixSuffixModule,
     DaffFormLabelModule,
     DaffErrorMessageModule,
-    ReactiveFormsModule,
   ],
   declarations: [
     DaffSelectComponent,
