+++ libs/design/select/src/select/select.component.scss
